Addressed SME feedback
