+++ finish/gateway/src/main/java/io/openliberty/guides/gateway/client/InventoryGatewayClient.java
@@ -12,10 +12,9 @@
 // end::copyright[]
 package io.openliberty.guides.gateway.client;
 
-import java.util.concurrent.CompletionStage;
+import java.util.Properties;
 
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -23,31 +22,20 @@ import javax.ws.rs.core.MediaType;
 
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
-import io.openliberty.guides.models.Job;
-import io.openliberty.guides.models.JobResult;
-import io.openliberty.guides.models.Jobs;
+import io.openliberty.guides.models.InventoryList;
 
-@RegisterRestClient(baseUri = "http://job-service:9080")
-@Path("/jobs")
-public interface JobClient {
+@RegisterRestClient(baseUri = "http://inventory-service:9080")
+@Path("/inventory")
+public interface InventoryGatewayClient {
 
-    // tag::getJobs[]
     @GET
+    @Path("systems")
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<Jobs> getJobs();
-    // end::getJobs[]
+    public InventoryList getInventory();
 
-    // tag::getJob[]
     @GET
-    @Path("{jobId}")
+    @Path("systems/{hostname}")
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<JobResult> getJob(@PathParam("jobId") String jobId);
-    // end::getJob[]
-
-    // tag::createJob[]
-    @POST
-    @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<Job> createJob();
-    // end::createJob[]
+    public Properties getProperties(@PathParam("hostname") String hostname);
 
 }
