+++ finish/gateway/src/main/java/io/openliberty/guides/gateway/client/JobGatewayClient.java
@@ -10,11 +10,10 @@
  *     IBM Corporation - Initial implementation
  *******************************************************************************/
 // end::copyright[]
-package io.openliberty.guides.gateway;
+package io.openliberty.guides.gateway.client;
 
 import java.util.concurrent.CompletionStage;
 
-import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
@@ -22,48 +21,33 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.eclipse.microprofile.rest.client.inject.RestClient;
+import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
-import io.openliberty.guides.gateway.client.JobClient;
-import io.openliberty.guides.models.JobList;
 import io.openliberty.guides.models.Job;
 import io.openliberty.guides.models.JobResult;
+import io.openliberty.guides.models.Jobs;
 
+@RegisterRestClient(baseUri = "http://job-service:9080")
 @Path("/jobs")
-public class JobResource {
+public interface JobGatewayClient {
 
-    @Inject
-    @RestClient
-    private JobClient jobClient;
-
+    // tag::getJobs[]
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<JobList> getJobs() {
-        return jobClient
-            .getJobs()
-            // tag::thenApplyAsync[]
-            .thenApplyAsync((jobs) -> {
-                return new JobList(jobs.getResults());
-            })
-            // end::thenApplyAsync[]
-            // tag::exceptionally[]
-            .exceptionally((ex) -> {
-                // Respond with empty list on error
-                return new JobList();
-            });
-            // end::exceptionally[]
-    }
+    public CompletionStage<Jobs> getJobs();
+    // end::getJobs[]
 
+    // tag::getJob[]
     @GET
     @Path("{jobId}")
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<JobResult> getJob(@PathParam("jobId") String jobId) {
-        return jobClient.getJob(jobId);
-    }
+    public CompletionStage<JobResult> getJob(@PathParam("jobId") String jobId);
+    // end::getJob[]
 
+    // tag::createJob[]
     @POST
     @Produces(MediaType.APPLICATION_JSON)
-    public CompletionStage<Job> createJob() {
-        return jobClient.createJob();
-    }
+    public CompletionStage<Job> createJob();
+    // end::createJob[]
+
 }
